pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "larapy"
version = "1.0.0"
description = "Laravel concepts implemented in Python Flask"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Larapy Team", email = "team@larapy.dev"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: Flask",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "Flask>=2.0.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.812",
]

# [project.scripts]
# larapy = "larapy.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/larapy"
Repository = "https://github.com/yourusername/larapy.git"
Documentation = "https://larapy.readthedocs.io"
"Bug Tracker" = "https://github.com/yourusername/larapy/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["larapy*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
